This is the REAL final codes!
It’s amazing! Last “Final codes” has problems, but it run well!
In line 67, I used if to do the counting, and it should not work, but
it did!
Okay, I C. It will not show any problem if I just check the button, and
not check the nScope graph!